+++ internal/gameplay/levels/definitions/level03.go
@@ -6,25 +6,25 @@ func Level03() LevelConfig {
 		Maze: MazeConfig{
 			Cols:                  12,
 			Rows:                  8,
-			DeadlyCells:           2,
+			LethalCells:           2,
 			FreezingCells:         4,
 			Patrollers:            4,
 			ExtraConnectionChance: 0.07,
 		},
 		Player: PlayerConfig{
-			Size: 12,
+			Size: DefaultPlayerSize,
 		},
 		Exit: ExitConfig{
 			Position: Coordinate{
 				X: 11,
 				Y: 7,
 			},
-			Size: 16,
+			Size: DefaultExitSize,
 		},
 		Collectibles: Collectibles{
 			Number: 4,
-			Size:   8,
-			Value:  1,
+			Size:   DefaultCollectibleSize,
+			Value:  DefaultCollectibleValue,
 		},
 		Timer: 60,
 	}